Update to v3.1

+++ apps/shell/shell-base64.c
@@ -28,7 +28,6 @@
  *
  * This file is part of the Contiki operating system.
  *
- * $Id: shell-base64.c,v 1.1 2010/02/03 20:37:29 adamdunkels Exp $
  */
 
 /**
@@ -87,7 +86,7 @@ base64_decode_char(char c)
 static int
 base64_add_char(struct base64_decoder_state *s, char c)
 {
-  if(isspace(c)) {
+  if(isspace((int)c)) {
     return 0;
   }
